Cool.
I wanted to harp on David Beazley's the function video. I decided to write my own version of it in the rust programming language. It was super fun but took me down quite an interesting rabbit hole. I plan to continue on with various implementations of sleep. I figure it will give me some other practice with the illustratious std::time::Duration sets.